\name{timezone-package}

\alias{timezone-package}

\alias{timezone}

\docType{package}

\title{
A single helper function for translating geographical coordinates into a TZ timezone name (Area/Location format).
}

\description{
This package primarily relies on the shapefile of the TZ timezones of the world created and maintained by Eric Muller and publicly available at \url{http://efele.net/maps/tz/world/} under a CC Zero license. Additional functionality exists to alternatively query the Google Maps API.
}

\author{
Colin Rundel <rundel@gmail.com>

Maintainer: Colin Rundel <rundel@gmail.com>
}

\keyword{ package }
